projects/capstone/completed_code/backend/requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
Flask-Cors
Flask-Migrate
flask_script
python-jose-cryptodome
psycopg2-binary
python-dotenv
gunicorn